과제 2차 완성

+++ server/controllers/posts/delete.js
@@ -1,15 +1,19 @@
 const { post } = require("../../models");
-const { isAuthorized } = require("../tokenFunctions");
+const { isAuthorized } = require("../../utils/tokenFunctions");
 module.exports = async (req, res) => {
   const id = req.params.id;
   const userInfo = isAuthorized(req);
 
   //유저가 로그인 상태가 아닐때
   if (!userInfo) {
-    return res.status(401).json({ message: "login plz" });
+    return res.status(401).json({ message: "Unauthorized request" });
   }
   try {
     let postInfo = await post.findOne({ where: { id } });
+    //삭제 요청한 게시물이 없을 때
+    if (!postInfo) {
+      return res.status(404).json({ message: "not exist" });
+    }
     //게시물 작성자와 요청자가 일치할 때 포스트 삭제
     if (userInfo.id === postInfo.userId) {
       await post.destroy({ where: { id } });
